+++ CHANGELOG.md
@@ -14,11 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 **Changed**
 - [Breaking] "Settings" added in last update were renamed to "options" due to unwieldy name. All methods and references were changed to reflect new name.
 - [Breaking] `forged` parameter was renamed to `forge_target` everywhere. `Forge#forged` was renamed to `#forge_target` accordingly.
-- [Breaking] Methods that raised `ArgumentError` on type mismatch now raise `TypeError`.
-- `Crucible` will now raise new `CircularAttributeDependencyError` error when detecting a cycle in attribute dependencies.
+- Methods that raised `ArgumentError` on class mismatch now raise `TypeError`.
+- `Forge` and `Sequence` will now raise new `ObjectInterfaceError` (inheriting from `TypeError`) when passed objects that don't implement the required interface.
+- `Crucible` will now raise new `CircularAttributeDependencyError` when detecting a cycle in attribute dependencies.
 
 **Fixed**
-- Calling `Forge#forge` with an unknown trait name now raises a clear error.
+- Calling `Forge#forge` with an unknown trait name now raises a clear `ArgumentError`.
 
 [Compare v0.3.0...main](https://github.com/trinistr/object_forge/compare/v0.3.0...main)